Move shuffle format metadata to Java module
