+++ sql-plugin-format/src/main/java/com/nvidia/spark/rapids/format/BufferTransferResponse.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright (c) 2026, NVIDIA CORPORATION.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 // automatically generated by the FlatBuffers compiler, do not modify
 
 package com.nvidia.spark.rapids.format;
@@ -50,4 +66,3 @@ public final class BufferTransferResponse extends Table {
     return o;
   }
 }
-
