+++ sql-plugin-format/src/main/java/com/nvidia/spark/rapids/format/TransferState.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright (c) 2026, NVIDIA CORPORATION.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+// automatically generated by the FlatBuffers compiler, do not modify
+
+package com.nvidia.spark.rapids.format;
+
+public final class TransferState {
+  private TransferState() { }
+  /**
+   * UCX transfer initiated on sender-side
+   */
+  public static final byte STARTED = 0;
+  /**
+   * Data has been compressed and requires meta update
+   */
+  public static final byte BUFFER_META_UPDATED = 1;
+
+  public static final String[] names = { "STARTED", "BUFFER_META_UPDATED", };
+
+  public static String name(int e) { return names[e]; }
+}
