--- sql-plugin/src/main/java/com/nvidia/spark/rapids/format/TransferState.java
@@ -1,20 +1,0 @@
-// automatically generated by the FlatBuffers compiler, do not modify
-
-package com.nvidia.spark.rapids.format;
-
-public final class TransferState {
-  private TransferState() { }
-  /**
-   * UCX transfer initiated on sender-side
-   */
-  public static final byte STARTED = 0;
-  /**
-   * Data has been compressed and requires meta update
-   */
-  public static final byte BUFFER_META_UPDATED = 1;
-
-  public static final String[] names = { "STARTED", "BUFFER_META_UPDATED", };
-
-  public static String name(int e) { return names[e]; }
-}
-
